requirements.txt
pyTelegramBotAPI==4.14.0
python-dotenv==1.1.1